refactor: Move `XpubDeserialize` to utils
